# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Authentication and Authorization API",
        "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).For this sample, you can use the api key `special-key` to test the authorization     filters.",
        "termsOfService": "http://Lanthanion.com",
        "license": {
            "name": "Apache 2.0",
            "url": "http://springdoc.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/api/chat/getparticularchat": {
            "post": {
                "tags": [
                    "chat-controller"
                ],
                "operationId": "getParticularRole",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GetParticularChatRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralChatResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/permissions/getallpermissions": {
            "post": {
                "tags": [
                    "permissions-controller"
                ],
                "operationId": "getAllRoles",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GeneralPagedRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralPagedPermissionsResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/roles/getparticularrole": {
            "post": {
                "tags": [
                    "roles-controller"
                ],
                "operationId": "getParticularRole_1",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GetParticularRoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetParticularRoleResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/createrole": {
            "post": {
                "tags": [
                    "roles-controller"
                ],
                "operationId": "createRole",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CreateRoleResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/deleterole": {
            "delete": {
                "tags": [
                    "roles-controller"
                ],
                "operationId": "deleteRoles",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/DeleteRoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/getparticularroleusers": {
            "post": {
                "tags": [
                    "roles-controller"
                ],
                "operationId": "getParticularRoleUsers",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GetParticularRoleUsersRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralPagedUsersResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/getparticularuserroles": {
            "post": {
                "tags": [
                    "roles-controller"
                ],
                "operationId": "getParticularUserRoles",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GetParticularUserRolesRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralPagedRolesResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/updaterole": {
            "put": {
                "tags": [
                    "roles-controller"
                ],
                "operationId": "updateRole",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GetParticularRoleResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/getallroles": {
            "post": {
                "tags": [
                    "roles-controller"
                ],
                "operationId": "getAllRoles_1",
                "requestBody": {
                    "content": {
                        "*/*": {
                            "schema": {
                                "$ref": "#/components/schemas/GeneralPagedRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "default response",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralRolesResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "GeneralPagedRequest": {
                "required": [
                    "isDeleted",
                    "items",
                    "page"
                ],
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    }
                }
            },
...
            "Users": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNumber": {
                        "type": "string"
                    },
                    "isLoggedIn": {
                        "type": "boolean"
                    },
                    "userType": {
                        "type": "string",
                        "enum": [
                            "ADVERTISER",
                            "AD_OWNER",
                            "AFFILIATE",
                            "ADMIN"
                        ]
                    },
                    "isDeleted": {
                        "type": "boolean"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "updatedBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "GeneralResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "Chat": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "updatedBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "GetParticularChatRequest": {
                "required": [
                    "chatId"
                ],
                "type": "object",
                "properties": {
                    "chatId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "GeneralChatResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Chat"
                    }
                }
            },
...
            "Roles": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "updatedBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "Permissions": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    },
                    "deletedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "updatedBy": {
                        "$ref": "#/components/schemas/Users"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "GeneralPagedPermissionsResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Permissions"
                        }
                    }
                }
            },
...
            "GetParticularRoleRequest": {
                "required": [
                    "roleId"
                ],
                "type": "object",
                "properties": {
                    "roleId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "GetParticularRoleResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Roles"
                    }
                }
            },
            "CreateRoleRequest": {
                "required": [
                    "createdBy",
                    "name"
                ],
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "createdBy": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "CreateRoleResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Roles"
                    }
                }
            },
            "DeleteRoleRequest": {
                "required": [
                    "roleId"
                ],
                "type": "object",
                "properties": {
                    "roleId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "GetParticularRoleUsersRequest": {
                "required": [
                    "isDeleted",
                    "items",
                    "page",
                    "roleId"
                ],
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "items": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "roleId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "isDeleted": {
                        "type": "boolean"
                    }
                }
            },
            "GeneralPagedUsersResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Users"
                        }
                    }
                }
            },
            "GetParticularUserRolesRequest": {
                "required": [
                    "userId"
                ],
                "type": "object",
                "properties": {
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "GeneralPagedRolesResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Roles"
                        }
                    }
                }
            },
            "UpdateRoleRequest": {
                "required": [
                    "roleId",
                    "updatedBy"
                ],
                "type": "object",
                "properties": {
                    "roleId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "updatedBy": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
            "GeneralRolesResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Roles"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "API KEY / API SECRET": {
                "type": "apiKey",
                "scheme": "apiKey"
            }
        }
    }
}